Moved precise string measurement code into its own folder and put each class in its own file.
