+++ TJAPlayer3/TJAPlayer3.csproj
@@ -95,7 +95,10 @@
     <Reference Include="WindowsBase" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="Common\PreciseStringMeasurement\CPreciseStringMeasurer.cs" />
     <Compile Include="Common\Discord.cs" />
+    <Compile Include="Common\PreciseStringMeasurement\DirectBitmap.cs" />
+    <Compile Include="Common\PreciseStringMeasurement\MeasureStringPreciselyCacheKey.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <EmbeddedResource Include="Properties\Discord.resx">
       <Generator>ResXFileCodeGenerator</Generator>
